trying out embedding

+++ src/graphcnn/network.py
@@ -130,7 +130,7 @@ class GraphCNNNetwork(object):
 
         def process_decoder_input(target_data, batch_size):
             ending = tf.strided_slice(target_data, [0, 0], [batch_size, -1], [1, 1])
-            dec_input = tf.concat([tf.cast(tf.fill([batch_size, 1], 1),tf.int64), ending], 1)
+            dec_input = tf.concat([tf.cast(tf.fill([batch_size, 1], 1),tf.float32), ending], 1)
             return dec_input
 
         def make_cell(rnn_size,init_range=0.1):
